# Fix Servfail on malformed query