@tonymet 2021-09-03
I was able to repro on a common dns server `8.8.8.8`
```
 go run . -s 8.8.8.8 -c 2 -n 10 www.ucla.edu
Benchmarking 8.8.8.8:53 via udp with 2 concurrent requests 


Total requests:	 20 of 20 (100.0%)
DNS success codes:	0

DNS response codes
	SERVFAIL:	20

Time taken for tests:	 151.803024ms
Questions per second:	 131.7

DNS timings, 20 datapoints
	 min:		 13.631488ms
	 mean:		 14.942208ms
	 [+/-sd]:	 597.78µs
	 max:		 16.252927ms

DNS distribution, 20 datapoints

```