@csucu 2021-09-04
Recursion is false by default, i get no errors when running with -r flag. i.e `go run . -r -s 8.8.8.8 www.ucla.edu`