@tonymet 2021-09-04
ok thanks for explaining. i was able to get this behavior with dig as well. 

* From the user's point of view, I think recurse=true would be a better default (e.g. it's the default with dig). 
* From developer's point of view, perf seems better with this patch. 

What are your thoughts on both